+++ RevitAddin.DA.Tester/Revit/App.cs
@@ -8,6 +8,7 @@ namespace RevitAddin.DA.Tester.Revit
 {
     public class App : DesignApplication
     {
+        public static string AddInName { get; set; } = "none";
         public override void OnStartup()
         {
             Console.WriteLine("----------------------------------------");
@@ -17,6 +18,13 @@ namespace RevitAddin.DA.Tester.Revit
             Console.WriteLine("----------------------------------------");
             Console.WriteLine($"AddInName: \t{Application.ActiveAddInId?.GetAddInName()}");
             Console.WriteLine("----------------------------------------");
+            Application.ApplicationInitialized += Application_ApplicationInitialized;   
+        }
+
+        private void Application_ApplicationInitialized(object sender, Autodesk.Revit.DB.Events.ApplicationInitializedEventArgs e)
+        {
+            var application = (Application)sender;
+            AddInName = application.ActiveAddInId?.GetAddInName();
         }
 
         public override void OnShutdown()