+++ RevitAddin.DA.Tester/Services/DesignAutomationController.cs
@@ -39,6 +39,7 @@ namespace RevitAddin.DA.Tester.Services
             }
 
             Console.WriteLine($"UI:\t{UI.IsValid()}");
+            Console.WriteLine($"App.AddInName:\t{App.AddInName}");
             Console.WriteLine("----------------------------------------");
             Console.WriteLine($"Shape:\t{typeof(ricaun.Revit.DB.Shape.Colors).Assembly}");
             Console.WriteLine($"Shape Location:\t{typeof(ricaun.Revit.DB.Shape.Colors).Assembly.Location}");